QUEEN: Implement original music volume handling
